refactor: extract grid column auto-width mixin
